Experiment JSON file
This file contains summaries of many of the experiments done in the course of this project, including all the hyperparameter settings, peak & final scores, and wall-clock time.  The main details missing are the "100-episode-average" and CPU vs GPU; most of them are GPU.  The scaling seemed to be that the GPU was roughly twice as fast, though there may be a batch-size dependency.